src/data-mgt/python/batch-jobs/airqo/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
confluent-kafka==1.7.0
docutils==0.16
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-bigquery==2.13.0
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
libcst==0.3.17
lockfile==0.12.2
luigi==3.0.2
mypy-extensions==0.4.3
numpy==1.20.1
packaging==20.9
pandas==1.2.3
proto-plus==1.18.1
protobuf==3.15.6
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-daemon==2.3.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
